inet-service/pyproject.toml
[project]
name = "inet-service"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi>=0.122.0",
    "uvicorn>=0.38.0",
    "mysql-connector-python>=9.1.0",
    "faker>=38.2.0",
]
